Split the proxy app to web and proxy
