+++ web/templates/web/index.html
@@ -3,8 +3,8 @@
 <!DOCTYPE html>
 <html>
 	<head>
-		<link href="{% static 'proxy/css/bootstrap.min.css' %}" rel="stylesheet">
-		<script src="{% static 'proxy/js/jQuery.min.js' %}"></script>
+		<link href="{% static 'web/css/bootstrap.min.css' %}" rel="stylesheet">
+		<script src="{% static 'web/js/jQuery.min.js' %}"></script>
         <style type="text/css">
             th {
                  text-align:center;
@@ -14,7 +14,7 @@
 	</head>
 
     <div class="container-fluid">
-        <h1>Ads Data</h1>
+        <h1>Rewrite Rules</h1>
         <p><a id="toggleButton" class="btn btn-primary" data-toggle="collapse" onclick="toggleCollapse();" data-is-toggled="true">Expand All</a></p>
         <div id="accordion">
             <table class="table-hover" border="1" style="width:100%">
@@ -28,9 +28,9 @@
                 </tr>
                 {% for entry in entries %}
                 <tr>
-                    <td style="min-width:200px;width:200px">{{ entry.url }}</td>
+                    <td style="min-width:125px;width:125px">{{ entry.url }}</td>
                     <td style="min-width:200px;width:200px">{{ entry.headers }}</td>
-                    <td>
+                    <td style="min-width:200px;width:200px">
                         <div class="panel-heading" id="heading{{ forloop.counter }}">
                             <h4 class="panel-title">
                                 <a data-toggle="collapse" href="#collapse{{ forloop.counter }}" aria-expanded="true" aria-controls="collapse{{ forloop.counter }}">
@@ -45,7 +45,7 @@
                     <td style="min-width:175px;width:175px">{{ entry.created_date }}</td>
                     <td style="min-width:175px;width:175px">{{ entry.modified_date }}</td>
                     <td style="min-width:33px;width:33px;text-align:center">
-                        <form id="delete" action="{% url 'proxy:delete_id' %}" method="post" style="display:inline">
+                        <form id="delete" action="{% url 'web:delete_id' %}" method="post" style="display:inline">
                             {% csrf_token %}
                             <input type="text" name="id" value="{{ entry.id }}" class="hidden">
                             <button type="submit" class="btn-sm btn-default glyphicon glyphicon-remove"
@@ -74,5 +74,5 @@
         }
     </script>
 
-    <script src="{% static 'proxy/js/bootstrap.min.js' %}"></script>
+    <script src="{% static 'web/js/bootstrap.min.js' %}"></script>
 </html>
