+++ web/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import url
 
 from . import views
 
-app_name = 'proxy'
+app_name = 'web'
 
 urlpatterns = [
     url(r'^$', views.index, name='index'),
